refactor(providers): split what a fixture is from what a provider must do
Review pass over this PR's own diff.

`provider-contract.ts` had crossed a thousand lines by holding two things: how
a recorded fixture is read back, and what every provider must do with one. The
first moves to `fixture-recording.ts`, which knows no trust constraint; the
second stays, at 750 lines of cases.

Cut, because a suite that carries dead flexibility is the last place it
belongs: `forbiddenRun`, a second throwing runner beside the recorded one that
already fails loudly for an invocation no fixture recorded — every kind now
takes the recorded runner; `conductorDb`, which nothing seeded; `RecordedCli.heal`
and `CaseOptions.loginProbeSlug`, which nothing called; `fixedAnswer` and
`recordedBody`, two helpers with one file of callers between them.

`ProviderFixtures.hookSpool.events` was declared and never read. It now earns
its keep: the hook case spools each of the provider's own tokens too, so
"everything the hook sharpens still observes from the transcripts alone" is
asserted against a spool that holds something real as well as against one that
holds nothing.
